+++ modules/qq-srv.zsh
@@ -4,13 +4,40 @@
 # qq-srv
 #############################################################
 
+qq-srv-help() {
+  cat << END
+
+qq-srv
+-------
+The srv namespace provides commands for hosting local services
+such as web, ftp, smb and other services for data exfil or transfer.
+
+Commands
+--------
+qq-srv-web: hosts a python3 web server
+qq-srv-ftp: hosts a python3 ftp server
+qq-srv-smb: hosts an impacket smb server 
+qq-srv-tftp: starts the atftpd service
+qq-srv-smtp: hosts a python3 smtp server
+qq-srv-updog: hosts an updog web server
+
+qq-srv-web-auto: hosts a python3 web server in /srv, port as $1
+qq-srv-php-auto: hosts a php web server in /srv, port as $1
+qq-srv-ftp-auto: hosts a python3 ftp server in /srv
+
+
+END
+}
+
+
+
 qq-srv-web() print -z "sudo python3 -m http.server 80"
 qq-srv-ftp() print -z "sudo python3 -m pyftpdlib -p 21 -w"
 qq-srv-smb() print -z "sudo impacket-smbserver -smb2supp F ."
 qq-srv-tftp() print -z "sudo service atftpd start"
-qq-srv-smtp() print -z "sudo python -m smtpd -c DebuggingServer -n 0.0.0.0:25"
+qq-srv-smtp() print -z "sudo python3 -m smtpd -c DebuggingServer -n 0.0.0.0:25"
 
-qq-srv-web-auto() {
+qq-srv-web-srv() {
     if [ "$#" -eq  "1" ]
     then
         pushd /srv; sudo python3 -m http.server $1; popd;
@@ -20,7 +47,7 @@ qq-srv-web-auto() {
 }
 alias srv-web="qq-srv-web-auto"
 
-qq-srv-php-auto() {
+qq-srv-php-srv() {
     if [ "$#" -eq  "1" ]
     then
         pushd /srv; sudo php -S 0.0.0.0:$1; popd;
@@ -30,7 +57,7 @@ qq-srv-php-auto() {
 }
 alias srv-php="qq-srv-php-auto"
 
-qq-srv-ftp-auto() {
+qq-srv-ftp-srv() {
     pushd /srv
     sudo python3 -m pyftpdlib -p 21 -w
     popd
@@ -41,7 +68,7 @@ qq-srv-updog() {
     print -z "updog -p 443 --ssl -p $(__rand 10)"
 }
 
-qq-srv-updog-auto() {
+qq-srv-updog-srv() {
     updog -p 443 --ssl -d /srv
 }
 alias srv-up